requirements.txt
pandas
numpy
scikit-learn
lightgbm
mlflow
shap
joblib
pyyaml
fastapi
uvicorn
streamlit
requests
sqlalchemy>=2.0
pytest
prometheus-client
python-dotenv
